Implemented authorization permissions to admin for portfolios and blogs

+++ app/helpers/application_helper_temp.rb
@@ -1,10 +1,10 @@
-module ApplicationHelper
+#module ApplicationHelper     *************   temp file when refactoring current_user   ***********
   def login_helper
-     if current_user.is_a?(GuestUser)
+    if current_user.is_a?(User)
+      link_to "Logout", destroy_user_session_path, method: :delete, data: { confirm: 'Are you sure?' }, class: "btn btn-sm btn-primary"
+    else
       (link_to "Register", new_user_registration_path, class: "btn btn-sm btn-primary") + " " +
       (link_to "Sign_in", new_user_session_path, class: "btn btn-sm btn-primary")
-     else
-      link_to "Logout", destroy_user_session_path, method: :delete, class: "btn btn-sm btn-primary"
     end
   end
